Add extension points to the dispatch code
