+++ gwtp-core/gwtp-dispatch-rest/src/main/java/com/gwtplatform/dispatch/rest/client/RestRequestBuilderFactory.java
@@ -16,127 +16,10 @@
 
 package com.gwtplatform.dispatch.rest.client;
 
-import java.util.List;
-import java.util.Map;
-
-import javax.inject.Inject;
-
-import org.jboss.errai.enterprise.client.jaxrs.JacksonTransformer;
-import org.jboss.errai.marshalling.client.Marshalling;
-
-import com.google.common.collect.Maps;
 import com.google.gwt.http.client.RequestBuilder;
-import com.google.gwt.safehtml.shared.UriUtils;
-import com.gwtplatform.dispatch.rest.shared.HttpMethod;
-import com.gwtplatform.dispatch.rest.shared.MetadataType;
 import com.gwtplatform.dispatch.rest.shared.RestAction;
-import com.gwtplatform.dispatch.rest.shared.RestParameter;
 import com.gwtplatform.dispatch.shared.ActionException;
 
-import static com.google.gwt.user.client.rpc.RpcRequestBuilder.MODULE_BASE_HEADER;
-
-public class RestRequestBuilderFactory {
-    private static final Map<HttpMethod, RequestBuilder.Method> HTTP_METHODS = Maps.newEnumMap(HttpMethod.class);
-    private static final String CONTENT_TYPE = "Content-Type";
-    private static final String JSON_UTF8 = "application/json; charset=utf-8";
-
-    static {
-        HTTP_METHODS.put(HttpMethod.GET, RequestBuilder.GET);
-        HTTP_METHODS.put(HttpMethod.POST, RequestBuilder.POST);
-        HTTP_METHODS.put(HttpMethod.PUT, RequestBuilder.PUT);
-        HTTP_METHODS.put(HttpMethod.DELETE, RequestBuilder.DELETE);
-        HTTP_METHODS.put(HttpMethod.HEAD, RequestBuilder.HEAD);
-    }
-
-    private final ActionMetadataProvider metadataProvider;
-    private final String baseUrl;
-    private final String securityHeaderName;
-
-    @Inject
-    RestRequestBuilderFactory(ActionMetadataProvider metadataProvider,
-                              @RestApplicationPath String baseUrl,
-                              @XCSRFHeaderName String securityHeaderName) {
-        this.metadataProvider = metadataProvider;
-        this.baseUrl = baseUrl;
-        this.securityHeaderName = securityHeaderName;
-    }
-
-    public <A extends RestAction<?>> RequestBuilder build(A action, String securityToken) throws ActionException {
-        RequestBuilder.Method httpMethod = HTTP_METHODS.get(action.getHttpMethod());
-        String url = buildUrl(action);
-
-        RequestBuilder requestBuilder = new RequestBuilder(httpMethod, url);
-
-        for (RestParameter param : action.getHeaderParams()) {
-            requestBuilder.setHeader(param.getName(), encode(param));
-        }
-
-        requestBuilder.setHeader(CONTENT_TYPE, JSON_UTF8);
-        requestBuilder.setHeader(MODULE_BASE_HEADER, baseUrl);
-
-        if (action.hasFormParams()) {
-            requestBuilder.setRequestData(buildQueryString(action.getFormParams()));
-        } else if (action.hasBodyParam()) {
-            requestBuilder.setRequestData(getSerializedValue(action, action.getBodyParam()));
-        }
-
-        if (securityToken != null && securityToken.length() > 0) {
-            requestBuilder.setHeader(securityHeaderName, securityToken);
-        }
-
-        return requestBuilder;
-    }
-
-    private String buildUrl(RestAction<?> restAction) throws ActionException {
-        String queryString = buildQueryString(restAction.getQueryParams());
-
-        if (!queryString.isEmpty()) {
-            queryString = "?" + queryString;
-        }
-
-        String path = buildPath(restAction.getPath(), restAction.getPathParams());
-
-        return baseUrl + path + queryString;
-    }
-
-    private String buildPath(String rawPath, List<RestParameter> params) throws ActionException {
-        String path = rawPath;
-
-        for (RestParameter param : params) {
-            path = path.replace("{" + param.getName() + "}", encode(param));
-        }
-
-        return path;
-    }
-
-    private String buildQueryString(List<RestParameter> params) throws ActionException {
-        StringBuilder queryString = new StringBuilder();
-
-        for (RestParameter param : params) {
-            queryString.append("&")
-                       .append(param.getName())
-                       .append("=")
-                       .append(encode(param));
-        }
-
-        if (queryString.length() != 0) {
-            queryString.deleteCharAt(0);
-        }
-
-        return queryString.toString();
-    }
-
-    private String encode(RestParameter value) throws ActionException {
-        return UriUtils.encode(value.getStringValue());
-    }
-
-    private String getSerializedValue(RestAction<?> action, Object object) throws ActionException {
-        Class<?> bodyClass = (Class<?>) metadataProvider.getValue(action, MetadataType.BODY_CLASS);
-
-        if (bodyClass == null || !Marshalling.canHandle(bodyClass)) {
-            throw new ActionException("Unable to serialize request body. No serializer found.");
-        } else {
-            return JacksonTransformer.toJackson(Marshalling.toJSON(object));
-        }
-    }
+public interface RestRequestBuilderFactory {
+    <A extends RestAction<?>> RequestBuilder build(A action, String securityToken) throws ActionException;
 }
