+++ gwtp-core/gwtp-dispatch-rpc-client/src/main/java/com/gwtplatform/dispatch/rpc/client/gin/PhoneGapDispatchAsyncModule.java
@@ -20,6 +20,7 @@ import javax.inject.Singleton;
 
 import com.google.gwt.inject.client.AbstractGinModule;
 import com.google.inject.Provides;
+import com.gwtplatform.dispatch.rpc.client.DefaultRpcDispatchCallFactory;
 import com.gwtplatform.dispatch.rpc.client.PhoneGapDispatchAsync;
 import com.gwtplatform.dispatch.rpc.client.RpcDispatchCallFactory;
 import com.gwtplatform.dispatch.rpc.shared.DispatchAsync;
@@ -29,7 +30,7 @@ import com.gwtplatform.dispatch.rpc.shared.DispatchServiceAsync;
 public class PhoneGapDispatchAsyncModule extends AbstractGinModule {
     @Override
     protected void configure() {
-        bind(RpcDispatchCallFactory.class).in(Singleton.class);
+        bind(RpcDispatchCallFactory.class).to(DefaultRpcDispatchCallFactory.class).in(Singleton.class);
 
         bind(DispatchAsync.class).to(PhoneGapDispatchAsync.class).in(Singleton.class);
     }
